Module comm::arc

Fork of the arc module in the rust stdlib.

Only the changes are documented here. See the stdlib for the rest.

In contrast to the stdlib arc module, this one also supports Arc objects that contain trait objects.

Example

use comm::arc::{Arc, ArcTrait};

struct X {
    x: u8
}

trait Y {
    fn f(&self) -> u8;
}

impl Y for X {
    fn f(&self) -> u8 {
        self.x
    }
}

let arc_trait: ArcTrait<Y> = unsafe {
    let arc = Arc::new(X { x: 3 });
    arc.as_trait(&*arc as &(Y+'static))
};

assert_eq!(arc_trait.f(), 3);

Structs

Arc

ArcTrait
An atomically reference counted wrapper of a trait object.

Weak

WeakTrait
A weak pointer to an ArcTrait.