Module comm::client

Modules

messages

Structs

Client
A Client handles receiving, processing, relaying, etc. of messages in the communication network. It internalizes all the complex logic that is specific to messaging, and exposes a few commands view a TaskSender and events via Events.

Enums

Event
Events emitted to any listeners registered with register_event_listener. They represent various client-level events.

ScheduledTask
A task to be performed some time in the future

Task
A command for the Client to execute immediately.

Type Definitions

Events
A receiver for receiving events from a running Client. Similar in purpose to TaskSender. A running Client is consumed and cannot be queried via method calls. You must instead subscribe to events it emits

TaskSender
A sender for issuing commands to the Client. Once a client is run, it is consumed and methods cannot be called on it. A TaskSender as the asynchronous interface for controlling a running Client.