Enum comm::client::Event

pub enum Event {
    ReceivedTextMessage(TextMessage),
    SentTextMessage(TextMessage),
    ReceivedMessageAcknowledgement(MessageAcknowledgement),
    Shutdown,
    Started,
}

Events emitted to any listeners registered with register_event_listener. They represent various client-level events.

Variants
ReceivedTextMessage(TextMessage)

We've received a message that was addressed to us

SentTextMessage(TextMessage)

We've sent or relayed a message

ReceivedMessageAcknowledgement(MessageAcknowledgement)

We've received an acknowledgement for a message we sent

Shutdown

The client has shut down

TODO: It might be useful for Shutdown to contain the serialized state of the entire Client, Network, RoutingTable, Buckets and Nodes so that a user may store it and later re-initialize a Client with this state in order to prevent re-bootstrapping from scratch.

Started

Network has started running. This does not necessary guarantee that we are "connected" to the network, because we haven't necessarily sent/received any messages. It just indicates that the Network is open for business.