Enum comm::client::ScheduledTask

pub enum ScheduledTask {
    DeliverMessage(Address, TextMessage),
}

A task to be performed some time in the future

Variants
DeliverMessage(Address, TextMessage)

Deliver a message to an address. This can be a message we are sending to someone else, or just a message we're relaying.