Struct comm::client::Client

pub struct Client { /* fields omitted */ }

A Client handles receiving, processing, relaying, etc. of messages in the communication network. It internalizes all the complex logic that is specific to messaging, and exposes a few commands view a TaskSender and events via Events.

Methods

impl Client

pub fn new(address: Address) -> Client

Creates a new Client. address is the address of this client, i.e. where other clients should send messages intended for this client.

pub fn run(self, network: Network) -> TaskSender

Starts a Client in its own thread and returns its task sender. This method consumes the Client, so all event listeners must be registered first.

pub fn register_event_listener(&mut self, event_listener: Sender<Event>)

Registers an event listener that should be sent every Event the client emits.