Struct comm::mpmc::bounded::Channel

pub struct Channel<T: Send> {
    // some fields omitted
}

An endpoint of a bounded MPMC channel.

Methods

impl<T: Send> Channel<T>

fn new(cap: usize) -> Channel<T>

Creates a new bounded MPMC channel with capacity at least cap.

Panic

Panics under any of the following conditions:

sizeof(usize) == 4 && cap > 2^15,

sizeof(usize) == 8 && cap > 2^31,

next_power_of_two(cap) * sizeof(T) >= isize::MAX.

fn send_sync(&self, val: T) -> Result<(), (T, Error)>

Sends a message over the channel. Blocks if the channel is full.

Error

Deadlock - All other endpoints are currently blocked trying to send a message.

fn send_async(&self, val: T) -> Result<(), (T, Error)>

Sends a message over the channel. Does not block if the channel is full.

Error

Full - The buffer is full.

fn recv_sync(&self) -> Result<T, Error>

Receives a message from the channel. Blocks if the channel is empty.

Error

Deadlock - All other endpoints are currently blocked trying to receive a message.

fn recv_async(&self) -> Result<T, Error>

Receives a message over the channel. Does not block if the channel is empty.

Error

Empty - The buffer is empty.