Function comm::mpsc::bounded_fast::new

pub unsafe fn new<'a, T: Sendable + 'a>(cap: usize) -> (Producer<'a, T>, Consumer<'a, T>)

Creates a new bounded MPSC channel with capacity at least cap.

Safety

This is unsafe because under just the right circumstances this implementation can lead to undefined behavior. Note that these circumstances are extremely rare and almost impossible on 64 bit systems.