Struct comm::mpsc::unbounded::Consumer

pub struct Consumer<T: Send> {
    // some fields omitted
}

The consuming end of an unbounded MPSC channel.

Methods

impl<T: Send> Consumer<T>

fn recv_sync(&self) -> Result<T, Error>

Receives a message from this channel. Blocks if the channel is empty.

Error

Disconnected - The channel is empty and all senders have disconnected.

fn recv_async(&self) -> Result<T, Error>

Receives a message from this channel. Does not block if the channel is empty.

Error

Disconnected - The channel is empty and all senders have disconnected.

Empty - The channel is empty.