Struct comm::select::Select

pub struct Select<'a> {
    // some fields omitted
}

Container for all targets being selected on.

Methods

impl<'a> Select<'a>

fn new() -> Select<'a>

Creates a new Select object.

fn add<T: Selectable<'a> + 'a>(&self, sel: &T)

Adds a target to the select object.

fn remove<T: Selectable<'a>>(&self, sel: &T) -> bool

Removes a target from the Select object. Returns true if the target was previously registered in the Select object, false otherwise.

fn wait<'b>(&self, ready: &'b mut [usize]) -> &'b mut [usize]

Waits for any of the targets in the Select object to become ready. The ids of the ready targets will be stored in ready. Returns the prefix containing the set of stored ids.

If the select object is empty, an empty slice is returned immediately.

fn wait_timeout<'b>(&self, ready: &'b mut [usize], duration: Option<Duration>) -> Option<&'b mut [usize]>

Waits for any of the targets in the Select object to become ready. The semantics are as for the wait function except that

if duration is some duration, then it will sleep for at most duration, and

if duration is none, then it will only check if at the time of the call any targets are ready and return immediately.

Return value

Returns None if the timeout expired.