Struct comm::select::WaitQueue

pub struct WaitQueue {
    // some fields omitted
}

A structure stored by Selectable objects to interact with Select objects that want to be notified when the Selectable object becomes ready.

Methods

impl WaitQueue

fn new() -> WaitQueue

Creates a new WaitQueue. This function does not allocate.

fn set_id(&mut self, id: usize)

Sets the id of the Selectable object containing the WaitQueue. This id must be the id returned by Selectable::as_selectable().unique_id(). This function must be called with the correct id before any other functions are called.

fn add(&mut self, load: Payload) -> usize

Add a Select object to the WaitQueue. Returns the number of Select objects contained in the WaitQueue after this call.

fn remove(&mut self, id: usize) -> usize

Removes a Select object from the WaitQueue. Returns the number of Select objects contained in the WaitQueue after this call.

fn notify(&mut self) -> usize

Notifies all Select objects contained in this WaitQueue that the Selectable object has become ready. Returns the number of Select objects contained in the WaitQueue after this call. This function might remove Select objects from the WaitQueue.

fn clear(&mut self)

Removes all Select objects from this wait queue and signals them that the Selectable object will no longer be available. This happens automatically when the WaitQueue is dropped.