Trait comm::select::_Selectable

pub unsafe trait _Selectable: Sync + Send {
    fn ready(&self) -> bool;
    fn register(&self, Payload);
    fn unregister(&self, id: usize);
}

The object that will be stored in a Select structure while the Selectable object is registered.

Implementing this trait is unsafe because the behavior is undefined if the implementation doesn't follow the rules below.

Required Methods

fn ready(&self) -> bool

Returns true if the object is ready, false otherwise.

This function must not try to acquire any locks that are also held while the implementation interacts with the WaitQueue object.

fn register(&self, Payload)

Registers a Select object with the Selectable object. The payload must be passed to the WaitQueue.

fn unregister(&self, id: usize)

Unregisters a Select objects from the Selectable object. The id must be passed to the WaitQueue.