Struct comm::spmc::bounded_fast::Consumer

pub struct Consumer<'a, T: Sendable + 'a> {
    // some fields omitted
}

A consumer of a bounded SPMC channel.

Methods

impl<'a, T: Sendable + 'a> Consumer<'a, T>

fn recv_sync(&self) -> Result<T, Error>

Receives a message from the channel. Blocks if the channel is empty.

Error

Disconnected - The sender has disconnected and the channel is empty.

fn recv_async(&self) -> Result<T, Error>

Receives a message over the channel. Does not block if the channel is empty.

Error

Disconnected - The sender has disconnected and the channel is empty.

Empty - The buffer is empty.