Module comm::spsc::ring_buf

A bounded SPSC channel that overwrites older messages when the buffer is full.

Example

Consider the case of an audio producer and consumer. If, at some point, the consumer is slow, you might not want to block the producer and instead overwrite older, unconsumed audio samples so that the delay between producer and consumer is bounded above by the buffer size of the channel.

Structs

Consumer
The sending half of an SPSC channel.

Producer
The producing half of an SPSC ring buffer channel.

Functions

new
Creates a new SPSC ring buffer channel.