Module comm::spsc::unbounded

An unbounded SPSC channel.

This channel can store an unbounded number of messages, however, sending a message might be more expensive than with bounded channels, and if the receiver can't dequeue messages fast enough, one might run out of memory.

Example

Consider the case of a short-lived producer. The producer might produce an unknown number of messages in a short amount of time and then disconnect. With an unbounded channel the producer will never block and the consumer can start processing the messages before the producer is finished.

Structs

Consumer
The consuming half on an unbounded SPSC channel.

Producer
The producing half on an unbounded SPSC channel.

Functions

new
Creates a new unbounded SPSC channel.